commitlint: resolve strict build errors
The error TS2802 is thrown by TypeScript when it encounters a
construct that can only be iterated over if the we enable
--downlevelIteration or if the --target is set to es2015
or higher. This is because older versions of JavaScript
(ES5 and below) do not support iteration over certain types.

With Array.from we can create an array from the Set<unknown>
and then iterate over the array. This should work without
needing to change any compiler flags however this solution
might not be suitable for all use cases, especially if the
Set<unknown> is large, as it requires creating an additional
array.

Also `"\g"` is a flag that indicates a global search, replacing
all matches in the string, not just the first one.